% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flippR.R
\name{flippr}
\alias{flippr}
\title{Transposes data frame through wide-to-long pivoting}
\usage{
flippr(df, main_column, stack_columns, stacked_new_name, last_column_name)
}
\arguments{
\item{df}{A data frame.}

\item{main_column}{A numeric or character vector.}

\item{stack_columns}{A numeric or character vector.}

\item{stacked_new_name}{A character vector.}

\item{last_column_name}{A character vector.}
}
\value{
A data frame.
}
\description{
Transposes data frame through wide-to-long pivoting
}
\examples{
flippr(iris, 5, c(1:4), "Attribute", "Length_Width")

flippr(iris, "Species", c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"), "Plant_Attribute", "Measurement")
}
