% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnownInfoIndivs.R
\name{KnownInfoIndivs}
\alias{KnownInfoIndivs}
\title{KnownInfoIndivs}
\usage{
KnownInfoIndivs(Your_Data)
}
\arguments{
\item{Your_Data}{The data for the individuals of interest, in Life History data-format}
}
\value{
Dataframe of individuals from Your_Data who have a known mother, sex, and birth date
}
\description{
Filter individuals to include only those for whom there is a known mother, sex, and birth date
}
