% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhosGrandma.R
\name{WhosGrandma}
\alias{WhosGrandma}
\title{WhosGrandma}
\usage{
WhosGrandma(Your_Data, Full_Data)
}
\arguments{
\item{Your_Data}{The data for the individuals of interest, in Life History data-format}

\item{Full_Data}{The data for the full known population in the study area, in Life History data-format}
}
\value{
A dataframe in Life History data-format of the life history data of the grandmothers of input individuals
}
\description{
Identify the grandmother of each individual of interest and return her life history information. Should be run only after HasGrandma to ensure all individuals do have grandmothers in the population first.
}
\examples{
MyData = HasGrandma(Your_Data, Full_Data) 
WhosGrandma(MyData, Full_Data)
}
